feat(api) : Add datacube tenders API
A very simple and flat API, with a basic but efficient authentication
scheme.

At the time of writing, there are about ~5000 tenders in the database.

With the companies attached, there are only ~700 left.

We'll see about time filtering later as the request to retrieve the
whole list takes about 200ms.